refactor: modularização de componentes do backend
