+++ backend/src/modules/apresentacao/apresentacao.module.ts
@@ -2,7 +2,5 @@ import { Module } from '@nestjs/common';
 
 import { ApresentacaoController } from './apresentacao.controller';
 
-@Module({
-    controllers: [ApresentacaoController]
-})
+@Module({ controllers: [ApresentacaoController] })
 export class ApresentacaoModule { }
