+++ backend/src/modules/auth/auth.module.ts
@@ -1,12 +1,12 @@
 import { PassportModule } from '@nestjs/passport';
+import { ConfigModule } from '@nestjs/config';
 import { Module } from '@nestjs/common';
 
+import { UsuarioModule } from '../usuario/usuario.module';
 import { GoogleStrategy } from './google.strategy';
 import { AuthController } from './auth.controller';
 import { AuthService } from './oauth.service';
 import { JwtStrategy } from './jwt.strategy';
-import { UsuarioModule } from '../usuario/usuario.module';
-import { ConfigModule } from '@nestjs/config';
 
 @Module({
   controllers: [AuthController],
