+++ backend/src/modules/auth/oauth.service.ts
@@ -1,9 +1,9 @@
 import { Injectable, InternalServerErrorException } from '@nestjs/common';
 import { sign } from 'jsonwebtoken';
 
-import * as jwt from '../credentials/secret-app.json';
+import * as jwt from '../../credentials/secret-app.json';
 import { UserGooglePayload } from 'src/interfaces/user-google-pay-load';
-import { UsuarioService } from '../services/usuario.service';
+import { UsuarioService } from '../../services/usuario.service';
 
 export enum Provider { GOOGLE = 'google' }
